+++ sdk/netapp/azure-mgmt-netapp/generated_samples/region_infos_list.py
@@ -38,6 +38,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: specification/netapp/resource-manager/Microsoft.NetApp/preview/2025-01-01-preview/examples/RegionInfos_List.json
+# x-ms-original-file: specification/netapp/resource-manager/Microsoft.NetApp/stable/2025-03-01/examples/RegionInfos_List.json
 if __name__ == "__main__":
     main()